requirements.txt
scrapy~=2.5.0
awesome-slugify~=1.6.5
html2markdown~=0.1.7
pickledb~=0.9.2
requests~=2.25.1